feat/#76: hook 분리
